requirements.txt
spacy
scikit-learn
sentence-transformers
pdfminer.six
python-docx
streamlit
PyMuPDF
